Moving compute resources into version dir
