+++ openstack/compute/v2/flavors/client.go
