+++ openstack/compute/v2/flavors/flavors.go
