+++ openstack/compute/v2/images/client.go
