+++ openstack/compute/v2/images/images_test.go
